+++ yash-cli/CHANGELOG.md
@@ -9,6 +9,14 @@ used by other programs.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.2] - Unreleased
+
+### Fixed
+
+- The shell no longer crashes when the `xtrace` option is enabled and the `PS4`
+  variable contains a command substitution, which previously caused infinitely
+  recursive expansion of `PS4`.
+
 ## [3.0.1] - 2025-10-13
 
 ### Changed
@@ -278,6 +286,7 @@ later.
 
 - Initial release of the shell
 
+[3.0.2]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.2
 [3.0.1]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.1
 [3.0.0]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-3.0.0
 [0.4.5]: https://github.com/magicant/yash-rs/releases/tag/yash-cli-0.4.5